Tweak action links on show pages

+++ app/helpers/channels_helper.rb
@@ -1,15 +1,18 @@
 module ChannelsHelper
-  def channel_action_links channel
+  def channel_action_links(channel, actions=[:update, :destroy])
     links = []
     type = channel.class.model_name.singular
-    if policy(channel).update?
+
+    if actions.include?(:update) && policy(channel).update?
       links << (link_to fa_icon('pencil', data: {toggle: 'tooltip', placement: 'top'}, title: 'Edit'),
         edit_channel_path(channel), class: 'btn btn-primary btn-xs action-button')
     end
-    if policy(channel).destroy?
+
+    if actions.include?(:destroy) && policy(channel).destroy?
       links << (link_to fa_icon('trash', data: {toggle: 'tooltip', placement: 'top'}, title: 'Delete'),
         channel, class: 'btn btn-danger btn-xs action-button', method: :delete)
     end
+
     links.join(' ').html_safe
   end
